VGA Mode Y (チェーン4 off、4プレーン) — DOS 版 DOOM のタイトル画面が正しく出る
- dev/chip/vga.rs: シーケンサ (0x3C4/5) と GC (0x3CE/F)。マップマスクでプレーンへ書き、
  読み出しマップのプレーンを RAM の窓に写して読み出しフック無し。CRTC 開始番地で合成
- 0xA0000 の窓は**チェーン4 が切れている間だけ**装置 (write8 / write_wide / fast_write /
  rep movs / JIT の脱出窓)。普段の mode 13h と Linux はこれまでどおり素通り
- framebuffer() は Cow: チェーン4 なら窓そのまま、Mode Y なら合成の一枚
- E2E: DOOM のタイトルが横に 4 枚並んでいない (行の周期 80 を見る)。DOOM_SHOT で撮影

+++ core/src/cpu/twobyte.rs
@@ -207,7 +207,11 @@ pub(crate) fn step_0f(m: &mut Machine, d: &Decoder, start_ip: u32) {
                 }
                 if lar {
                     // アクセス権: 上位ダブルワードの bit 8-23 (16bit形は bit 8-15 だけ)
-                    Some(if d.opsize32 { hi & 0x00FF_FF00 } else { hi & 0xFF00 })
+                    Some(if d.opsize32 {
+                        hi & 0x00FF_FF00
+                    } else {
+                        hi & 0xFF00
+                    })
                 } else {
                     let mut limit = (lo & 0xFFFF) | (hi & 0x000F_0000);
                     if hi & 0x0080_0000 != 0 {
@@ -218,7 +222,7 @@ pub(crate) fn step_0f(m: &mut Machine, d: &Decoder, start_ip: u32) {
             };
             match result {
                 Some(v) => {
-                    m.cpu.set_reg_w(reg as usize, v, d.opsize32);
+                    m.cpu.set_reg_w(reg, v, d.opsize32);
                     m.cpu.set_flag(super::ZF, true);
                 }
                 None => m.cpu.set_flag(super::ZF, false),
